# [branch-2.1] Optimize schema change (#5265)