--- a/be/src/common/config.h
+++ b/be/src/common/config.h
@@ -694,6 +694,9 @@ CONF_Int64(send_rpc_runtime_filter_timeout_ms, "1000");
 
 CONF_Int32(max_batch_publish_latency_ms, "100");
 
+// enable optimized implementation of schema change
+CONF_Bool(enable_schema_change_v2, "true");
+
 } // namespace config
 
 } // namespace starrocks