--- a/be/src/storage/column_mapping.h
+++ b/be/src/storage/column_mapping.h
@@ -32,6 +32,10 @@ struct ColumnMapping {
     // <0: use default value
     // >=0: use origin column
     int32_t ref_column{-1};
+
+    // The base reader selects part of the column
+    // so it's different to ref_column
+    int32_t ref_base_reader_column_index = -1;
     // normally for default value. stores values for filters
     WrapperField* default_value{nullptr};
     // materialize view transform function used in schema change