--- a/be/src/storage/vectorized/memtable.cpp
+++ b/be/src/storage/vectorized/memtable.cpp
@@ -32,7 +32,7 @@ MemTable::MemTable(int64_t tablet_id, const TabletSchema* tablet_schema, const s
           _rowset_writer(rowset_writer),
           _aggregator(nullptr),
           _mem_tracker(mem_tracker) {
-    _vectorized_schema = ChunkHelper::convert_schema_to_format_v2(*tablet_schema);
+    _vectorized_schema = std::move(ChunkHelper::convert_schema_to_format_v2(*tablet_schema));
     if (_keys_type == KeysType::PRIMARY_KEYS && _slot_descs->back()->col_name() == LOAD_OP_COLUMN) {
         // load slots have __op field, so add to _vectorized_schema
         auto op_column = std::make_shared<starrocks::vectorized::Field>((ColumnId)-1, LOAD_OP_COLUMN,
@@ -42,6 +42,26 @@ MemTable::MemTable(int64_t tablet_id, const TabletSchema* tablet_schema, const s
         _has_op_slot = true;
     }
 
+    if (_keys_type != KeysType::DUP_KEYS) {
+        // The ChunkAggregator used by MemTable may be used to aggregate into a large Chunk,
+        // which is not suitable for obtaining Chunk from ColumnPool,
+        // otherwise it will take up a lot of memory and may not be released.
+        _aggregator = std::make_unique<ChunkAggregator>(&_vectorized_schema, 0, INT_MAX, 0);
+    }
+}
+
+MemTable::MemTable(int64_t tablet_id, const Schema& schema, RowsetWriter* rowset_writer, int64_t max_buffer_size,
+                   MemTracker* mem_tracker)
+        : _tablet_id(tablet_id),
+          _vectorized_schema(std::move(schema)),
+          _tablet_schema(nullptr),
+          _slot_descs(nullptr),
+          _keys_type(schema.keys_type()),
+          _rowset_writer(rowset_writer),
+          _aggregator(nullptr),
+          _use_slot_desc(false),
+          _max_buffer_size(max_buffer_size),
+          _mem_tracker(mem_tracker) {
     if (_keys_type != KeysType::DUP_KEYS) {
         // The ChunkAggregator used by MemTable may be used to aggregate into a large Chunk,
         // which is not suitable for obtaining Chunk from ColumnPool,
@@ -78,7 +98,7 @@ size_t MemTable::write_buffer_size() const {
 }
 
 bool MemTable::is_full() const {
-    return write_buffer_size() >= config::write_buffer_size;
+    return write_buffer_size() >= _max_buffer_size;
 }
 
 bool MemTable::insert(const Chunk& chunk, const uint32_t* indexes, uint32_t from, uint32_t size) {
@@ -87,14 +107,22 @@ bool MemTable::insert(const Chunk& chunk, const uint32_t* indexes, uint32_t from
     }
 
     size_t cur_row_count = _chunk->num_rows();
-    // For schema change, FE will construct a shadow column.
-    // The shadow column is not exist in _vectorized_schema
-    // So the chunk can only be accessed by the subscript
-    // instead of the column name.
-    for (int i = 0; i < _slot_descs->size(); ++i) {
-        const ColumnPtr& src = chunk.get_column_by_slot_id((*_slot_descs)[i]->id());
-        ColumnPtr& dest = _chunk->get_column_by_index(i);
-        dest->append_selective(*src, indexes, from, size);
+    if (_use_slot_desc) {
+        // For schema change, FE will construct a shadow column.
+        // The shadow column is not exist in _vectorized_schema
+        // So the chunk can only be accessed by the subscript
+        // instead of the column name.
+        for (int i = 0; i < _slot_descs->size(); ++i) {
+            const ColumnPtr& src = chunk.get_column_by_slot_id((*_slot_descs)[i]->id());
+            ColumnPtr& dest = _chunk->get_column_by_index(i);
+            dest->append_selective(*src, indexes, from, size);
+        }
+    } else {
+        for (int i = 0; i < _vectorized_schema.num_fields(); i++) {
+            const ColumnPtr& src = chunk.get_column_by_index(i);
+            ColumnPtr& dest = _chunk->get_column_by_index(i);
+            dest->append_selective(*src, indexes, from, size);
+        }
     }
 
     if (chunk.has_rows()) {
@@ -257,7 +285,7 @@ void MemTable::_sort(bool is_final) {
     for (uint32_t i = 0; i < _chunk->num_rows(); ++i) {
         _permutations[i] = {i, i};
     }
-    if (_tablet_schema->num_key_columns() <= 3) {
+    if (_vectorized_schema.num_key_fields() <= 3 && _use_slot_desc) {
         _sort_chunk_by_columns();
     } else {
         _sort_chunk_by_rows();
@@ -487,7 +515,7 @@ private:
 };
 
 void MemTable::_sort_chunk_by_columns() {
-    for (int i = _tablet_schema->num_key_columns() - 1; i >= 0; --i) {
+    for (int i = _vectorized_schema.num_key_fields() - 1; i >= 0; --i) {
         Column* column = _chunk->get_column_by_index(i).get();
         PrimitiveType slot_type = (*_slot_descs)[i]->type().type;
         if (column->is_nullable()) {
@@ -531,7 +559,7 @@ void MemTable::_sort_chunk_by_columns() {
 void MemTable::_sort_chunk_by_rows() {
     pdqsort(false, _permutations.begin(), _permutations.end(),
             [this](const MemTable::PermutationItem& l, const MemTable::PermutationItem& r) {
-                size_t col_number = _tablet_schema->num_key_columns();
+                size_t col_number = _vectorized_schema.num_key_fields();
                 int compare_result = 0;
                 for (size_t col_index = 0; col_index < col_number; ++col_index) {
                     const auto& left_col = _chunk->get_column_by_index(col_index);