--- a/be/test/storage/vectorized/schema_change_test.cpp
+++ b/be/test/storage/vectorized/schema_change_test.cpp
@@ -515,6 +515,12 @@ TEST_F(SchemaChangeTest, schema_change_with_sorting) {
     column_mapping->ref_column = 3;
     column_mapping->ref_base_reader_column_index = 3;
     indexs->emplace_back(3);
+    column_mapping = chunk_changer.get_mutable_column_mapping(4);
+    column_mapping->ref_column = -1;
+    const TabletColumn& hll_column_schema = new_tablet->tablet_schema().column(4);
+    column_mapping->default_value = WrapperField::create(hll_column_schema);
+    ASSERT_TRUE(column_mapping->default_value != nullptr);
+    column_mapping->default_value->from_string("");
 
     _sc_procedure = new (std::nothrow) SchemaChangeWithSorting(
             &chunk_changer, config::memory_limitation_per_thread_for_schema_change * 1024 * 1024 * 1024);