The main optimizations include:
 1. Only use the required columns when schema-changing
 2. Use memtable for sorted schema change
 3. More accurate memory usage control

The performance will be improved mostly when the materialized view has only a few columns.
The boost can reach 27 times.
|case|optimize|main|boost|remark|
|---|---|---|---|---|
|sql1|2452ms|66217ms|27x|select two integer column as mv|
|sql2|7253ms|85697ms|11.8x|select ten integer column as mv|
|sql3|18413ms|78477ms|4.3x|select two string column as mv|
|sql4|48653ms|139069ms|2.9x|select ten string column as mv|
|sql5|480785ms|682649ms|1.42x|reorder all column memory_limitation_per_thread_for_schema_change=100|
|sql5|554830ms|OOM|N/A|memory_limitation_per_thread_for_schema_change=10|

```
sql1: create materialized view v1 as select comment_id, position from github_events
sql2: create materialized view v2 as select comment_id, position, line, number, comments, review_comments, commits, additions, deletions, push_size from github_events
sql3: create materialized view v3 as select action, title from github_events
sql4: create materialized view v4 as select action, title, event_type, repo_name, actor_login, state, path, head_sha, commit_id, release_name from github_events
sql5: create materialized view v5 as select title, created_at, event_type, repo_name, file_time, actor_login, updated_at, action, comment_id, body, path, position, line, ref, ref_type, creator_user_login, number, labels, state, locked, assignee, assignees, comments, author_association, closed_at, merged_at, merge_commit_sha, requested_reviewers, requested_teams, head_ref, head_sha, base_ref, base_sha, merged, mergeable, rebaseable, mergeable_state, merged_by, review_comments, maintainer_can_modify, commits, additions, deletions, changed_files, diff_hunk, original_position, commit_id, original_commit_id, push_size, push_distinct_size, member_login, release_tag_name, release_name review_state from github_events order by title, created_at
```

## What type of PR is this：
- [ ] bug
- [ ] feature
- [ ] enhancement
- [ ] refactor
- [ ] others

## Which issues of this PR fixes ：
<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

## Problem Summary(Required) ：
<!-- (Please describe the changes you have made. In which scenarios will this bug be triggered and what measures have you taken to fix the bug?) -->