@Astralidea 2022-06-21
run starrocks_be_unittest